Move settings singleton into its own folder and update it to match how we use other singletons.
