+++ packages/salesforcedx-vscode-core/src/settings/index.ts
@@ -1,0 +1,10 @@
+/*
+ * Copyright (c) 2017, salesforce.com, inc.
+ * All rights reserved.
+ * Licensed under the BSD 3-Clause license.
+ * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+import { SfdxCoreSettings } from './sfdxCoreSettings';
+
+export const sfdxCoreSettings = SfdxCoreSettings.getInstance();
